rename lib to components, add hooks dir, fix order object so its always the correct type and not a mishmash of limit vs market order
