+++ libs/react-helpers/src/components/deal-ticket/type-selector.tsx
@@ -1,21 +1,21 @@
-import { Order, OrderSide } from '.';
 import { Button } from '@vegaprotocol/ui-toolkit';
+import { Order, OrderType } from '../../hooks/use-order-state';
 
-interface SideSelectorProps {
+interface TypeSelectorProps {
   order: Order;
-  onSelect: (side: OrderSide) => void;
+  onSelect: (type: OrderType) => void;
 }
 
-export const SideSelector = ({ order, onSelect }: SideSelectorProps) => {
+export const TypeSelector = ({ order, onSelect }: TypeSelectorProps) => {
   return (
     <div className="flex gap-8 mb-20">
-      {Object.entries(OrderSide).map(([key, value]) => {
+      {Object.entries(OrderType).map(([key, value]) => {
         return (
           <Button
             onClick={() => onSelect(value)}
             className="flex-1"
             type="button"
-            variant={value === order.side ? 'accent' : undefined}
+            variant={value === order.type ? 'accent' : undefined}
             key={key}
           >
             {key}
