+++ libs/react-helpers/src/components/vega-wallet/connectors/injected-connector.ts
