+++ libs/react-helpers/src/components/vega-wallet/connectors/rest-connector.ts
