+++ libs/react-helpers/src/index.ts
@@ -1,3 +1,3 @@
-export * from './lib/react-helpers';
-export * from './lib/trading';
-export * from './lib/vega-wallet';
+export * from './components/vega-wallet';
+export * from './components/trading';
+export * from './components/deal-ticket';
