--- libs/react-helpers/src/lib/react-helpers.spec.tsx
@@ -1,10 +1,0 @@
-import { render } from '@testing-library/react';
-
-import ReactHelpers from './react-helpers';
-
-describe('ReactHelpers', () => {
-  it('should render successfully', () => {
-    const { baseElement } = render(<ReactHelpers />);
-    expect(baseElement).toBeTruthy();
-  });
-});
