--- libs/react-helpers/src/lib/react-helpers.tsx
@@ -1,12 +1,0 @@
-/* eslint-disable-next-line */
-export interface ReactHelpersProps {}
-
-export function ReactHelpers(props: ReactHelpersProps) {
-  return (
-    <div>
-      <h1>Welcome to ReactHelpers!</h1>
-    </div>
-  );
-}
-
-export default ReactHelpers;
